show mem cpu usage

+++ src/sdk/galaxy.h
@@ -64,8 +64,8 @@ struct NodeDescription {
     int32_t task_num;
     double cpu_share;
     int32_t mem_share;
-    int32_t mem_used;
-    int32_t cpu_used;
+    int32_t  mem_used;
+    double cpu_used;
 };
 
 struct JobInstanceDescription : JobDescription{
